104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,16 +5,16 @@
  */
 int main(void)
 {
-	long int a = 1, b = 2, c, count;
+	unsigned long long int a = 1, b = 2, c, count;
 
-	printf("%ld, %ld", a, b);
+	printf("%llu, %llu", a, b);
 
 	for (count = 3; count <= 98; count++)
 	{
 		c = a + b;
 		a = b;
 		b = c;
-		printf(", %ld", c);
+		printf(", %llu", c);
 	}
 
 	printf("\n");
